+++ apps/www/src/app/(main)/(docs)/components/otp-input/examples/radius.tsx
@@ -1,0 +1,24 @@
+import { OtpInput, type OtpInputProps } from "@jamsr-ui/react";
+
+export const OtpInputRadius = () => {
+  const radii: OtpInputProps["radius"][] = [
+    "none",
+    "sm",
+    "md",
+    "lg",
+    "xl",
+    "2xl",
+    "3xl",
+  ];
+  return (
+    <div className="grid gap-4">
+      {radii.map((radius) => (
+        <OtpInput
+          key={radius}
+          radius={radius}
+          label="Enter your OTP"
+        />
+      ))}
+    </div>
+  );
+};