+++ apps/www/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web
 
+## 0.1.187
+
+### Patch Changes
+
+- menu, drawer..
+  - @jamsr-ui/react@0.0.205
+
 ## 0.1.186
 
 ### Patch Changes