+++ packages/components/confirmation/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/confirmation
 
+## 0.0.70
+
+### Patch Changes
+
+- menu, drawer..
+
 ## 0.0.69
 
 ### Patch Changes