+++ packages/components/menu/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/menu
 
+## 0.0.95
+
+### Patch Changes
+
+- menu, drawer..
+
 ## 0.0.94
 
 ### Patch Changes