+++ packages/components/otp-input/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/otp-input
 
+## 0.0.74
+
+### Patch Changes
+
+- menu, drawer..
+
 ## 0.0.73
 
 ### Patch Changes